% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantm_stats.R
\name{quantm_mm}
\alias{quantm_mm}
\title{A CLDL Quantm Function}
\usage{
quantm_mm(measure, df, spill_values)
}
\arguments{
\item{measure}{the measure for the lmer, as STRING (likely "resid")}

\item{df}{the dataframe, contains a column named spill, subjCode, objCode, NPSubj, NPObj, and subject}
}
\value{
writes model comparison ouput to external file
}
\description{
This function generates maximal models and writes model comparisons to a file for 1 DATAFRAME
}
\examples{
quantm_mm("resid", fprtRaw)
quantm_mm("resid", rpd)
lapply(c(fprtRaw,fprt), function(x) quantm_mm("resid",x))

}

