% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantm_stats.R
\name{quantm_resid}
\alias{quantm_resid}
\title{A CLDL Quantm Function}
\usage{
quantm_resid(factor, df)
}
\arguments{
\item{factor}{the factor as a STRING, corresponds to a named column in the dataframe}

\item{df}{the dataframe, contains a column named}
}
\value{
returns 2 dataframes to the global environment; one with log resid and one with raw resid
}
\description{
This function creates dataframe and log/raw resids for 1 factor
}
\examples{
quantm_resid("FPRT", qdata)
quantm_resid("RPD", qdata)

}

