% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantm_stats.R
\name{quantm_stats}
\alias{quantm_stats}
\title{A CLDL Quantm Function}
\usage{
quantm_stats(factor, df, spill_values)
}
\arguments{
\item{factor}{the factor as STRING, must correspond to a column in the dataframe}

\item{df}{the dataframe, contains a column named with the factor, spill, subjCode, objCode, NPsubj, NPobj, and subject}
}
\value{
writes model comparison ouput to external files for factorRaw and factor(log)
}
\description{
This function combines calls to quantm_resid and quantm_mm to run all stats on 1 factor (this fxn calls mm on both log and raw datasets separately for the factor)
}
\examples{
quantm_stats("FPRT", qdata, c(0,1,2))
lapply(c("FPRT", "RPD", "TFT"), function(x) quantm_stats(x, qdata, c(0,1,2)))

}

